+++ odoo/local-src/specific_sale/models/__init__.py
@@ -6,3 +6,4 @@ from . import res_partner
 from . import sale_subscription
 from . import subscription_template
 from . import account_analytic_account
+from . import stock_picking