+++ odoo/local-src/specific_sale/wizard/__init__.py
@@ -1,2 +1,3 @@
 from . import sale_refusal
 from . import mrp_invoicing
+from . import picking_update_delivery_date